requirements.txt
Django==1.5.1
Scrapy==0.18.0
Twisted==13.1.0
Whoosh==2.5.2
distribute==0.6.31
dj-database-url==0.2.2
dj-static==0.0.5
django-haystack==2.1.0
django-toolbelt==0.0.1
gunicorn==17.5
html2text==3.200.3
lxml==3.2.3
psycopg2==2.5.1
pyOpenSSL==0.13
pysolr==3.1.0
queuelib==1.0
requests==1.2.3
solr==0.1
static==0.4
w3lib==1.3
wsgiref==0.1.2
zope.interface==4.0.5
